requirements.txt
customtkinter>=5.2.0
pywifi>=1.1.12
requests>=2.31.0
python-dotenv>=1.0.0 